requirements.txt
customtkinter
mysql-connector-python
python-dotenv
secure-smtplib